+++ src/presentation/graphql/schema/index.ts
@@ -19,4 +19,5 @@ export const typeDefs: string[] = [
   loadSchema('base.graphql'),
   loadSchema('accounts.graphql'),
   loadSchema('budgets.graphql'),
+  loadSchema('categories.graphql'),
 ];